add timeout

+++ Time_functions.h
@@ -7,8 +7,8 @@
 #include <string_view>
 #include <optional>
 
-std::optional<int> stringViewToInt(std::string_view str) {
-    int result;
+std::optional<int64_t> stringViewToInt(std::string_view str) {
+    int64_t result;
     auto [ptr, ec] = std::from_chars(str.data(), str.data() + str.size(), result);
     if (ec == std::errc()) {
         return result;
