+++ enums_typedef.h
@@ -9,5 +9,5 @@ enum class side {BUY, SELL, NONE};
 typedef int64_t Price;
 typedef int32_t Shares;
 typedef uint64_t Timestamp;
-typedef int32_t OrderId;
+typedef int64_t OrderId;
 typedef int32_t SymbolId;
